Separate out standalone init subcommand
