+++ pkg/crinit/standalone/standalone.go
@@ -14,38 +14,17 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package standalone contains the implementation of the `standalone` subcommand
-// of `crinit`, which deploys a cluster registry as a standalone API server in
-// a host Kubernetes cluster.
 package standalone
 
 import (
 	"io"
 
 	"k8s.io/api/core/v1"
-	client "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/cluster-registry/pkg/crinit/options"
-	"k8s.io/cluster-registry/pkg/crinit/util"
 
-	"github.com/golang/glog"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
-)
-
-var (
-	longInitCommandDescription = `
-	Init initializes a standalone cluster registry.
-
-	The standalone cluster registry is hosted inside a Kubernetes
-	cluster but handles its own authentication and authorization.
-	The host cluster must be specified using the
-        --host-cluster-context flag.`
-	initCommandExample = `
-	# Initialize a standalone cluster registry named foo
-	# in the host cluster whose local kubeconfig
-	# context is bar.
-	crinit standalone init foo --host-cluster-context=bar`
 )
 
 type standaloneClusterRegistryOptions struct {
@@ -65,109 +44,17 @@ func (o *standaloneClusterRegistryOptions) Bind(flags *pflag.FlagSet) {
 		"Enables token authentication for the cluster registry API server. Defaults to false.")
 }
 
-// NewCmdStandalone defines the `standalone` command that bootstraps a cluster registry
-// inside a host Kubernetes cluster.
-func NewCmdStandalone(cmdOut io.Writer, pathOptions *clientcmd.PathOptions, defaultServerImage, defaultEtcdImage string) *cobra.Command {
-	opts := &standaloneClusterRegistryOptions{}
-
+// NewCmdStandalone defines the `standalone` command with `init` and `delete`
+// subcommands to bootstrap or remove a cluster registry inside a host
+// Kubernetes cluster.
+func NewCmdStandalone(cmdOut io.Writer, pathOptions *clientcmd.PathOptions,
+	defaultServerImage, defaultEtcdImage string) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "standalone",
 		Short: "Subcommands to manage a standalone cluster registry",
 		Long:  "Commands used to manage a standalone cluster registry. That is, a cluster registry that is not aggregated with another Kubernetes API server.",
 	}
 
-	initCmd := &cobra.Command{
-		Use:     "init CLUSTER_REGISTRY_NAME --host-cluster-context=HOST_CONTEXT",
-		Short:   "Initialize a standalone cluster registry",
-		Long:    longInitCommandDescription,
-		Example: initCommandExample,
-		Run: func(cmd *cobra.Command, args []string) {
-			err := opts.SetName(args)
-			if err != nil {
-				glog.Fatalf("error: %v", err)
-			}
-
-			err = validateOptions(opts)
-			if err != nil {
-				glog.Fatalf("error: %v", err)
-			}
-
-			err = opts.MarshalOptions()
-			if err != nil {
-				glog.Fatalf("error: %v", err)
-			}
-
-			hostConfig, err := util.GetClientConfig(pathOptions, opts.Host, opts.Kubeconfig).ClientConfig()
-			if err != nil {
-				glog.Fatalf("error: %v", err)
-			}
-			hostClientset, err := client.NewForConfig(hostConfig)
-			if err != nil {
-				glog.Fatalf("error: %v", err)
-			}
-			err = Run(opts, cmdOut, hostClientset, pathOptions)
-			if err != nil {
-				glog.Fatalf("error: %v", err)
-			}
-		},
-	}
-
-	flags := initCmd.Flags()
-	opts.BindCommon(flags, defaultServerImage, defaultEtcdImage)
-	opts.Bind(flags)
-
-	cmd.AddCommand(initCmd)
+	cmd.AddCommand(newSubCmdInit(cmdOut, pathOptions, defaultServerImage, defaultEtcdImage))
 	return cmd
-}
-
-// validateOptions ensures that options are valid.
-func validateOptions(opts *standaloneClusterRegistryOptions) error {
-	opts.APIServerServiceType = v1.ServiceType(opts.apiServerServiceTypeString)
-	return opts.ValidateCommonOptions()
-}
-
-// Run initializes a cluster registry.
-func Run(opts *standaloneClusterRegistryOptions, cmdOut io.Writer,
-	hostClientset client.Interface, pathOptions *clientcmd.PathOptions) error {
-
-	err := opts.CreateNamespace(cmdOut, hostClientset)
-	if err != nil {
-		return err
-	}
-
-	svc, ips, hostnames, err := opts.CreateService(cmdOut, hostClientset)
-	if err != nil {
-		return err
-	}
-
-	credentials, err := opts.GenerateCredentials(cmdOut, svc.Name, ips, hostnames,
-		opts.apiServerEnableHTTPBasicAuth, opts.apiServerEnableTokenAuth)
-	if err != nil {
-		return err
-	}
-
-	err = opts.CreateAPIServerCredentialsSecret(hostClientset, credentials)
-	if err != nil {
-		return err
-	}
-
-	pvc, err := opts.CreatePVC(cmdOut, hostClientset, svc.Name)
-	if err != nil {
-		return err
-	}
-
-	err = opts.CreateAPIServer(cmdOut, hostClientset, opts.apiServerEnableHTTPBasicAuth,
-		opts.apiServerEnableTokenAuth, false, ips, pvc, "default")
-	if err != nil {
-		return err
-	}
-
-	err = opts.UpdateKubeconfig(cmdOut, pathOptions, svc, ips, hostnames,
-		credentials)
-	if err != nil {
-		return err
-	}
-
-	return opts.WaitForAPIServer(cmdOut, hostClientset, pathOptions, ips,
-		hostnames, svc)
 }
